compress all values instead of just zeros
